+++ src/features/blog/api/queries.ts
@@ -1,0 +1,8 @@
+export {
+	getAllPosts,
+	getPostBySlug,
+	getPostsByCategory,
+	getPostsByTag,
+	getAllSlugs,
+	getAllCategories,
+} from "~/lib/sanity/queries";
