+++ src/features/home/index.ts
@@ -1,0 +1,6 @@
+export { default as Hero } from "./components/Hero";
+export { default as StatsBar } from "./components/StatsBar";
+export { default as Services } from "./components/Services";
+export { default as Industries } from "./components/Industries";
+export { default as WhyBobatech } from "./components/WhyBobatech";
+export { default as FAQ } from "./components/FAQ";
